refactor(admin): split UserManagementCard into focused sub-components
Extract RoleManagement and AccountActions from the 390-line
UserManagementCard into separate single-responsibility components.
Compute permission flags (canManage, canAssignRoles, callerRank)
once in UserDetailCards and pass as props, eliminating duplicated
hierarchy logic between parent and child.

Closes #293
